# JOSResampleReader performance and faster quality settings