--- a/src/respec/JOSResampleReader.cpp
+++ b/src/respec/JOSResampleReader.cpp
@@ -49,7 +49,7 @@ static inline int lrint_impl(double x)
 
 AUD_NAMESPACE_BEGIN
 
-JOSResampleReader::JOSResampleReader(std::shared_ptr<IReader> reader, SampleRate rate) :
+JOSResampleReader::JOSResampleReader(std::shared_ptr<IReader> reader, SampleRate rate, Quality quality) :
 	ResampleReader(reader, rate),
 	m_channels(CHANNELS_INVALID),
 	m_n(0),
@@ -57,6 +57,28 @@ JOSResampleReader::JOSResampleReader(std::shared_ptr<IReader> reader, SampleRate
 	m_cache_valid(0),
 	m_last_factor(0)
 {
+	switch(quality)
+	{
+	case QUALITY_LOW:
+		m_len = m_len_low;
+		m_L = m_L_low;
+		m_coeff = m_coeff_low;
+		break;
+	case QUALITY_MEDIUM:
+		m_len = m_len_medium;
+		m_L = m_L_medium;
+		m_coeff = m_coeff_medium;
+		break;
+	case QUALITY_HIGH:
+		m_len = m_len_high;
+		m_L = m_L_high;
+		m_coeff = m_coeff_high;
+		break;
+	default:
+		m_len = m_len_low;
+		m_L = m_L_low;
+		m_coeff = m_coeff_low;
+	}
 }
 
 void JOSResampleReader::reset()