Backstory: I was looking at what takes up time in Blender during rendering out an animation from a Sequencer. Audio resampling inside Audaspace was "huh feels somewhat costly?" impression.

Currently in Audaspace there's a choice between linear resampler (okay for preview, but not great for final mix), or "extremely high quality preset" for rendering the final mix; with nothing in between. Now, I don't know whether inside say Blender it should be a user preference (like what Audacity has) or a project setting. Based on my findings below, I can't really tell any difference between the current High JOS resampler and the newly added Medium JOS resampler setting, except that the medium one is quite a bit faster.

The filter parameters used:
- High (i.e. current, unchanged): Nz = 136, L = 2304, freq = 0.963904, Kaiser Window B = 16
- Medium: Nz = 42, L = 500, freq = 0.916636, Kaiser Window B = 12
- Low: Nz = 16, L = 128, freq = 0.834068, Kaiser Window B = 10

### Performance

Timings, rendering out 300 frames (frames 1500-1800) of Sprite Fright Edit blender studio data set. Whole rendering process takes about 15 seconds. If using the newly added Medium quality setting, instead of currently existing Audaspace code, the time taken by `JOSResampleReader::read` during rendering:
- Windows (Ryzen 5950X, VS2022): 1.18s -> 0.51s
- Mac (M1 Max, clang 15): 2.68s -> 1.07s

i.e. this change saves about 10% of total rendering time on a Mac.

### Spectrograms

Spectrograms of `Swept_float.wav` from https://src.infinitewave.ca/ in Audacity, using these visualization settings:

![jos-spectrograms](https://github.com/audaspace/audaspace/assets/348087/e88e0dd1-1388-4531-9a4c-b6fcc324c29d)

![jos-resampler](https://github.com/audaspace/audaspace/assets/348087/ae1c7824-5341-4eeb-ad49-cd3bd9b2caa6)