@aras-p 2023-12-31
> 1. What speed improvement do you get from the SSE2 optimization

You mean from SSE2 lrint? It's in the message of that commit. Resampling 3min30sec 44.1khz stereo wav file into 48khz one, on Windows / Ryzen 5950X: 7.92s -> 7.33s.

I guess rewriting whole function to be SIMD might make it yet faster, but that would be way more invasive change.

> 2. I am not a fan of copying the filters from libsamplerate and would rather have filters designed similarly to the one I created for audaspace

Makes sense, indeed I mostly have no idea what I'm doing :) How would I design medium/low quality filter using your script? It has three parameters, that's a large parameter space to explore for someone who has no idea about them.