@neXyon 2023-12-31
> > 1. What speed improvement do you get from the SSE2 optimization
> 
> You mean from SSE2 lrint? It's in the message of that commit. Resampling 3min30sec 44.1khz stereo wav file into 48khz one, on Windows / Ryzen 5950X: 7.92s -> 7.33s.
> 
> I guess rewriting whole function to be SIMD might make it yet faster, but that would be way more invasive change.

Ah, so about a 7.5% decrease in time. That may not sound much, but it's significant.

> > 2. I am not a fan of copying the filters from libsamplerate and would rather have filters designed similarly to the one I created for audaspace
> 
> Makes sense, indeed I mostly have no idea what I'm doing :) How would I design medium/low quality filter using your script? It has three parameters, that's a large parameter space to explore for someone who has no idea about them.

Well, you can see the parameters that libsamplerate used. `L` is the `coeffs` variable you had to give the code anyway, so that should have a similar value. The number of values in the array then mostly depends on `Nz * L`, so you get an approximate `Nz` by dividing the length of the array by `L`. Since the length of the array also determines the speed of your computations, a lower number will give you a faster results, but the quality will of course be poorer. And then for a given `L` and `Nz` you want to set the window parameter `B` in a sensible way. From [wikipedia](https://en.wikipedia.org/wiki/Kaiser_window):

> ... is a non-negative real number that determines the shape of the window. In the frequency domain, it determines the trade-off between main-lobe width and side lobe level, which is a central decision in window design.

The ideal filter would be 1 (= 0 dB) for 0 to 1 Hz and then drop to 0 (- inf dB), but that would require an infinitely long filter, so that's not achievable. So instead you want to stay at 0 dB as long as possible and then drop off as sharply as possible without ripples or a spike that goes above 0 dB and then stay as low as possible at below -1XX dB usually, that's often called the signal-to-noise ratio (SNR). `B` lets you trade off how long it stays at 0 dB/how quickly it goes down vs. how low it gets (and stays).

To get approximately something that libsamplerate uses, try these to start with:

- medium: `L = 500`, `Nz = 42`, `B = 12`
- low: `L = 128`, `Nz = 16`, `B = 10`

From here it's an art... Basically you you want to make sure that your passband is as nice as possible (again: flat at 0 as long as possible, no ripples and then going down as quickly as possible) given a specific length of the array and the noise floor dB you want. That's what the plots are trying to show you, since "Passband" and "Transition" are just zooming in on the relevant parts of the "Magnitude" plot. You can ignore the "Response" and "Phase" plots. The latter should be fine if you only vary the 3 parameters and don't try other filter design methods/windows/etc.