@neXyon 2024-01-02
Thanks! I see you were happy with my suggested values already ;)