@aras-p 2024-01-02
> I see you were happy with my suggested values already ;)

Yeah I played around with several neighboring parameter values but the filter graphs did not look better.